remove previous tslint comment before adding new ones
